+++ backend/core/src/applications/types/application-program-option.ts
@@ -1,3 +1,0 @@
-import { ApplicationPreferenceOption } from "./application-preference-option"
-
-export class ApplicationProgramOption extends ApplicationPreferenceOption {}