+++ backend/core/src/multiselect-question/dto/jurisdictionFilterTypeToFieldMap.ts
@@ -1,0 +1,6 @@
+import { MultiselectQuestionFilterKeys } from "./multiselect-question-filter-keys"
+
+export const jurisdictionFilterTypeToFieldMap: Record<MultiselectQuestionFilterKeys, string> = {
+  jurisdiction: "multiselectQuestionJurisdictions.id",
+  applicationSection: "multiselectQuestions.applicationSection",
+}