+++ backend/core/src/seeder/seeds/listings/listing-triton-seed.ts
@@ -80,8 +80,7 @@ const tritonListing: ListingSeedType = {
   leasingAgentOfficeHours: "Monday - Friday, 9:00 am - 5:00 pm",
   leasingAgentPhone: "650-437-2039",
   leasingAgentTitle: "Business Manager",
-  listingPreferences: [],
-  listingPrograms: [],
+  listingMultiselectQuestions: [],
   name: "Test: Triton",
   postmarkedApplicationsReceivedByDate: null,
   programRules: null,
@@ -220,10 +219,10 @@ export class ListingTritonSeed extends ListingDefaultSeed {
       name: "Test: Triton 2",
       publishedAt: dayjs(new Date()).subtract(2.5, "hour"),
       assets: getDefaultAssets(),
-      listingPreferences: [
+      listingMultiselectQuestions: [
         {
-          preference: await this.preferencesRepository.findOneOrFail({
-            title: getLiveWorkPreference(alamedaJurisdiction.name).title,
+          multiselectQuestion: await this.multiselectQuestionsRepository.findOneOrFail({
+            text: getLiveWorkPreference(alamedaJurisdiction.name).text,
           }),
           ordinal: 2,
         },