+++ sites/partners/pages/settings/index.tsx
@@ -1,5 +1,6 @@
 import React, { useContext, useState } from "react"
 import Head from "next/head"
+import { ApplicationSection } from "@bloom-housing/backend-core"
 import {
   AppearanceSizeType,
   AppearanceStyleType,
@@ -19,8 +20,8 @@ import dayjs from "dayjs"
 import { AuthContext } from "@bloom-housing/shared-helpers"
 import { faClone, faPenToSquare, faTrashCan } from "@fortawesome/free-regular-svg-icons"
 import Layout from "../../layouts"
-import { useJurisdictionalPreferenceList } from "../../lib/hooks"
 import PreferenceDrawer from "../../src/settings/PreferenceDrawer"
+import { useJurisdictionalMultiselectQuestionList } from "../../lib/hooks"
 
 const Settings = () => {
   const { profile } = useContext(AuthContext)
@@ -28,35 +29,36 @@ const Settings = () => {
   const [deleteModal, setDeleteModal] = useState(false)
   const [preferenceDrawerOpen, setPreferenceDrawerOpen] = useState(false)
 
-  const { data, loading } = useJurisdictionalPreferenceList(
+  const { data, loading } = useJurisdictionalMultiselectQuestionList(
     profile?.jurisdictions?.reduce((acc, curr) => {
       return `${acc}${","}${curr.id}`
-    }, "")
+    }, ""),
+    ApplicationSection.preferences
   )
 
   const iconContent = () => {
     return (
       <div className={"flex justify-end"}>
         <div className={"w-max"}>
-          <span onClick={() => setPreferenceDrawerOpen(true)} className={"cursor-pointer"}>
+          <button onClick={() => setPreferenceDrawerOpen(true)}>
             <Icon
               symbol={faPenToSquare}
               size={"medium"}
               fill={IconFillColors.primary}
               className={"mr-5"}
             />
-          </span>
-          <span onClick={() => alert("copy")} className={"cursor-pointer"}>
+          </button>
+          <button onClick={() => alert("copy")}>
             <Icon
               symbol={faClone}
               size={"medium"}
               fill={IconFillColors.primary}
               className={"mr-5"}
             />
-          </span>
-          <span onClick={() => setDeleteModal(true)} className={"cursor-pointer"}>
+          </button>
+          <button onClick={() => setDeleteModal(true)}>
             <Icon symbol={faTrashCan} size={"medium"} fill={IconFillColors.alert} />
-          </span>
+          </button>
         </div>
       </div>
     )
@@ -77,7 +79,7 @@ const Settings = () => {
             cellClassName={"px-5 py-3"}
             data={data?.map((preference) => {
               return {
-                name: { content: preference?.title },
+                name: { content: preference?.text },
                 jurisdiction: {
                   content: preference?.jurisdictions?.reduce((acc, item, index) => {
                     return `${acc}${index > 0 ? ", " : ""}${item.name}`