+++ sites/partners/src/listings/PaperListingForm/sections/SelectAndOrder.tsx
@@ -11,9 +11,9 @@ import {
   StandardTableData,
 } from "@bloom-housing/ui-components"
 import { useFormContext } from "react-hook-form"
-import { Preference, Program } from "@bloom-housing/backend-core/types"
+import { ApplicationSection, MultiselectQuestion } from "@bloom-housing/backend-core/types"
 
-type SelectAndOrderSection = Preference | Program
+type SelectAndOrderSection = MultiselectQuestion
 
 type SelectAndOrderProps = {
   listingData: SelectAndOrderSection[]
@@ -25,16 +25,19 @@ type SelectAndOrderProps = {
   drawerTitle: string
   drawerButtonText: string
   dataFetcher: (
-    jurisdiction?: string
+    jurisdiction?: string,
+    applicationSection?: ApplicationSection
   ) => {
     data: SelectAndOrderSection[]
     loading: boolean
     error: any
   }
   formKey: string
+  applicationSection: ApplicationSection
 }
 
 const SelectAndOrder = ({
+  applicationSection,
   listingData,
   setListingData,
   title,
@@ -67,7 +70,7 @@ const SelectAndOrder = ({
         fetchedData.map((item) => {
           setValue(
             `${formKey}.${item.id}`,
-            editedListingData.some((existingItem) => existingItem.title === item.title)
+            editedListingData.some((existingItem) => existingItem.text === item.text)
           )
         })
       }
@@ -79,7 +82,7 @@ const SelectAndOrder = ({
   const draggableTableData: StandardTableData = useMemo(
     () =>
       draftListingData.map((item) => ({
-        name: { content: item.title },
+        name: { content: item.text },
         action: {
           content: (
             <div className="flex">
@@ -104,7 +107,7 @@ const SelectAndOrder = ({
     () =>
       listingData.map((item, index) => ({
         order: { content: index + 1 },
-        name: { content: item.title },
+        name: { content: item.text },
         action: {
           content: (
             <div className="flex">
@@ -131,7 +134,7 @@ const SelectAndOrder = ({
       const newDragOrder = []
       dragOrder.forEach((item) => {
         newDragOrder.push(
-          draftListingData.filter((draftItem) => draftItem.title === item.name.content)[0]
+          draftListingData.filter((draftItem) => draftItem.text === item.name.content)[0]
         )
       })
       setDraftListingData(newDragOrder)
@@ -141,7 +144,7 @@ const SelectAndOrder = ({
 
   const jurisdiction: string = watch("jurisdiction.id")
 
-  const { data: fetchedData = [] } = dataFetcher(jurisdiction)
+  const { data: fetchedData = [] } = dataFetcher(jurisdiction, applicationSection)
 
   const formTableHeaders = {
     order: "t.order",
@@ -239,11 +242,11 @@ const SelectAndOrder = ({
                       id={`${formKey}.${item.id}`}
                       name={`${formKey}.${item.id}`}
                       type="checkbox"
-                      label={item.title}
+                      label={item.text}
                       register={register}
                       inputProps={{
                         defaultChecked: draftListingData.some(
-                          (existingItem) => existingItem.title === item.title
+                          (existingItem) => existingItem.text === item.text
                         ),
                       }}
                     />
@@ -252,26 +255,28 @@ const SelectAndOrder = ({
               })
             : t("listings.selectJurisdiction")}
         </div>
-        <Button
-          id="addPreferenceSaveButton"
-          type="button"
-          className={"mt-4"}
-          styleType={AppearanceStyleType.primary}
-          size={AppearanceSizeType.normal}
-          onClick={() => {
-            const formData = getValues()
-            const formItems = []
-            fetchedData.forEach((uniqueItem) => {
-              if (formData[formKey][uniqueItem.id]) {
-                formItems.push(uniqueItem)
-              }
-            })
-            setDraftListingData(formItems)
-            setSelectDrawer(null)
-          }}
-        >
-          {t("t.save")}
-        </Button>
+        {jurisdiction && (
+          <Button
+            id="addPreferenceSaveButton"
+            type="button"
+            className={"mt-4"}
+            styleType={AppearanceStyleType.primary}
+            size={AppearanceSizeType.normal}
+            onClick={() => {
+              const formData = getValues()
+              const formItems = []
+              fetchedData.forEach((uniqueItem) => {
+                if (formData[formKey] && formData[formKey][uniqueItem.id]) {
+                  formItems.push(uniqueItem)
+                }
+              })
+              setDraftListingData(formItems)
+              setSelectDrawer(null)
+            }}
+          >
+            {t("t.save")}
+          </Button>
+        )}
       </Drawer>
     </>
   )